Re-add mpi tests
